+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - ADD: Tracking for APRS Objects (has to be enabled per Object).
 - ADD: Splitter to resize Map and PoI details.
 - ADD: support for overlay image.
+- ADD: Simple PoI filter.
 
 ## v0.4.0